requirements.txt
# Physics-Informed MeshGraphNets Requirements

# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0

# NGSolve for mesh generation and FEM
ngsolve>=6.2.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
Pillow>=9.0.0

# Utilities
h5py>=3.6.0
tqdm>=4.62.0
opencv-python>=4.5.0